third_party/blink/web_tests/svg/dynamic-updates/SVGFilterPrimitiveStandardAttributes-dom-y-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' attribute of the SVGFilterPrimitiveStandardAttributes object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('y') is "100"
PASS gradientElement.getAttribute('y') is "0"
PASS successfullyParsed is true

TEST COMPLETE
